Day 11 - Connecting cables
As with the previous couple of days, the first part was super easy [for me] and then part 2 started out seeming easy and turned into me wasting 4+ hours trying to think of optimizations. I spent most of that time working on one particular optimization, but it doesn't seem to save anytime after all. That's disappointing, but I am moving on.